Expose createFilterFromQuery helper
