+++ client/lib/minimongo/bson.ts
@@ -1,29 +1,4 @@
-export const enum BSONType {
-	Double = 1,
-	String,
-	Object,
-	Array,
-	BinData,
-	/** @deprecated */
-	Undefined,
-	ObjectId,
-	Boolean,
-	Date,
-	Null,
-	Regex,
-	/** @deprecated */
-	DBPointer,
-	JavaScript,
-	/** @deprecated */
-	Symbol,
-	JavaScriptWithScope,
-	Int,
-	Timestamp,
-	Long,
-	Decimal,
-	MinKey = -1,
-	MaxKey = 127,
-}
+import { BSONType } from './types';
 
 export const getBSONType = <T>(v: T): BSONType => {
 	if (typeof v === 'number') {
