+++ client/lib/minimongo/index.ts
@@ -1,0 +1,3 @@
+import { compileDocumentSelector } from './query';
+
+export const createFilterFromQuery = compileDocumentSelector;
